Improved code runtime, etc
Improved the algorithm slightly (eg replacing long for loops with while loops and put more emphasis on counters) and fixed a few errors. There's a significant decrease in runtime -- I can actually run I=8 on my laptop without spending an eternity. Currently I is set to 5 because it runs very quickly and is a good number of approximations.